requirements.txt
disnake
requests
matplotlib
python-dotenv